% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{max2D_cpp}
\alias{max2D_cpp}
\title{Compute the test statistic for the robust independence test of Kolmogorov-Smirnov's type}
\usage{
max2D_cpp(x, y)
}
\arguments{
\item{x, y}{the two sequences of vector that are applied to the function.}
}
\description{
Compute the maximum distance between the joint empirical cumulative
distribution function (e.c.d.f.) and the product of the marginals e.c.d.f.
multiplied by square root of n
}
