% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spearmanCore}
\alias{spearmanCore}
\title{Compute a quantity used in the calculation of the Spearman correlation}
\usage{
spearmanCore(x, y)
}
\arguments{
\item{x, y}{the two sequences of vector that are applied to the function.}
}
\description{
This is a fast function implemented in Rcpp.
}
